requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-Bcrypt==0.7.1
Flask-Limiter==3.5.0
python-dotenv==1.0.0
requests==2.31.0
pytz==2023.3
Werkzeug==2.3.7
MarkupSafe==2.1.3
PyJWT==2.8.0
supabase==2.3.4
postgrest==0.13.2
